BVC Refactor: Moved URLBarDelegate outside bvc
